fix, feat:  add kubernetes_manifests library and rename the library file name to match the published library (#85) (#87)
* fix: rename the library file name to match the published library

The library published in Charmhub is actually called resource_dispatcher and not
kubernetes_manifests. This commit renames the file so it actually matches the published library.

Part of canonical/bundle-kubeflow#1209

* feat: add kubernetes_manifests library

This commit creates the kubernetes_manifests library with a new LIBID that
will correspond to the kubernetes_manifests library published in Charmhub.

Part of canonical/bundle-kubeflow#1209

Co-authored-by: Daniela Plascencia <daniela.plascencia@canonical.com>